final version

+++ src/main/java/utils/ResponseData.java
@@ -5,24 +5,24 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDateTime;
 
 public class ResponseData {
-    private long time;
+    private double time;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "HH:mm:ss dd-MM-yyyy")
     private LocalDateTime localTime;
     private boolean result;
     private String message;
 
-    public ResponseData(long time, LocalDateTime localTime, boolean result, String message) {
+    public ResponseData(double time, LocalDateTime localTime, boolean result, String message) {
         this.time = time;
         this.localTime = localTime;
         this.result = result;
         this.message = message;
     }
 
-    public long getTime() {
+    public double getTime() {
         return time;
     }
 
-    public void setTime(long time) {
+    public void setTime(double time) {
         this.time = time;
     }
